added exists method
Exists() is a fast way to (not) find an element and check for the
non-existence. Find and FindAll attempt retries on the premise that the
element _should_ be found. Exists() assumes the element wouldn't be
found (and quickly). You can always call Find and Assert.Throws or call
FindAll and check for an empty list if you require a thorough
non-existence check.