improvements and finally found an appropriate place for FindsBy support
PageFactory will initialize and cache page objects. A good use for
FindsBy is when you need multiple identifiers for a dynamic web element.
This should not be confused with two static web elements that look the
same in the AUT, are used interchangable (and predictably), and do not
have the same identifiers. These types of web elements should exist on
two separate page objects and do not require multiple identifiers since
they are, in fact, two distinct web elements used at different points in
the application.